+++ docs/CHANGES.rst
@@ -5,6 +5,9 @@ Changes
 2.5 (unreleased)
 ----------------
 
+- Add parameter to autocomplete to search with exact match [URB-2696]
+  [jchandelle]
+
 - Change MultiSelectionWidget to MultiSelect2Widget [URB-2575]
   [jchandelle]